:truck: Renaming all reference from New Category to Category detail
Classes and resources renamed to match the new structure once the
screen now handles updates as well
