+++ app/src/main/res/layout/fragment_category_new.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="viewModel"
-            type="com.escodro.alkaa.ui.category.create.NewCategoryViewModel" />
+            type="com.escodro.alkaa.ui.category.detail.CategoryDetailViewModel" />
 
     </data>
 
@@ -106,7 +106,7 @@
             android:layout_marginTop="16dp"
             android:background="@color/colorAccent"
             android:elevation="48dp"
-            android:text="@string/new_category_add"
+            android:text="@string/category_detail_add"
             android:textColor="@android:color/white"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
